move url path middleware to runtime layer
